requirements.txt
streamlit
requests
aiohttp
pandas
python-dotenv